feat: add tagged types example
PostgreSQL does not support arrays with objects of different types, but
this can be handled with C functions so that you have have an array of
internal objects that are tagged as different types.

It is then necessary to have a constructor that can construct tagged types as
well, so that is also added as part of the example.