feat: add C++ version of table scanner
To develop C++ support for PostgreSQL, we create a C++ version of the table
scanner and add features to support this.

This adds an iterator that provides a table scan iterator. It is an input
iterator and will provide the rows one by one. It is not possible to change
direction once the scanning started.