feat: memory resident view for ephemeral data
Implement a memory-resident view of data that can be updated using DSM module.
This allow you to keep a view of ephemeral data available for quick access from
other backend without requiring updates of tables.

It also adds trigger functions that can be used to create an updatable memory
view.